Enabled auth in calendar

+++ api/cal/api.py
@@ -5,14 +5,17 @@ import cal.schemas
 
 from typing import List
 
+from security.api import AuthBearer
+
 router = Router()
 
-@router.post('/post_record')
+@router.post('/post_record', auth=AuthBearer())
 def post_cal_record(request, r: cal.schemas.CalendarRecordIn):
     try:
         if r.note.__len__() < 4:
             raise ValueError('Note must contain at least 4 chars')
         record = r.get_object()
+        record.owner_token = request.auth.owner_token
         record.save()
     except Exception as e:
         print(f'Failed to post record: {e}')
@@ -21,13 +24,13 @@ def post_cal_record(request, r: cal.schemas.CalendarRecordIn):
         return {'result': 'success', 'code': 0}
 
 
-@router.get('/get_record', response=cal.schemas.CalendarRecordOut)
+@router.get('/get_record', response=cal.schemas.CalendarRecordOut, auth=AuthBearer())
 def get_cal_record(request, id: int):
     try:
         if id <= 0:
             raise ValueError('Bad id')
         rs = records.models.Record.objects.filter(
-            id=id, record_type=records.models.RecordTypes.CALENDAR)
+            id=id, record_type=records.models.RecordTypes.CALENDAR, owner_token=request.auth.owner_token)
         r = rs[0]
         return r
     except Exception as e:
@@ -35,19 +38,19 @@ def get_cal_record(request, id: int):
         return HttpResponseBadRequest(content=f'{e}')
 
 
-@router.get('/get_records', response=List[cal.schemas.CalendarRecordOut])
+@router.get('/get_records', response=List[cal.schemas.CalendarRecordOut], auth=AuthBearer())
 def get_cal_records(request, limit: int = 100, offset: int = 0):
     if limit > 500:
         limit = 500
     if offset < 0:
         offset = 0
-    return list(records.models.Record.objects.filter(record_type=records.models.RecordTypes.CALENDAR).order_by('-id')[offset:(offset+limit)])
+    return list(records.models.Record.objects.filter(record_type=records.models.RecordTypes.CALENDAR, owner_token=request.auth.owner_token).order_by('-id')[offset:(offset+limit)])
 
 
-@router.delete('/delete_record')
+@router.delete('/delete_record', auth=AuthBearer())
 def delete_cal_record(request, id: int):
     try:
-        ret = records.models.Record.objects.filter(id=id, record_type=records.models.RecordTypes.CALENDAR).delete()
+        ret = records.models.Record.objects.filter(id=id, record_type=records.models.RecordTypes.CALENDAR, owner_token=request.auth.owner_token).delete()
         if ret[0] == 0:
             raise ValueError('Deleted zero elements')
     except Exception as e:
@@ -56,12 +59,12 @@ def delete_cal_record(request, id: int):
     else:
         return {'result': 'success', 'code': 0}
 
-@router.patch('/update_deadline')
+@router.patch('/update_deadline', auth=AuthBearer())
 def update_deadline(request, record_id: int, new_deadline: cal.schemas.Deadline):
     try:
         if record_id <= 0:
             raise ValueError('Bad id')
-        ret = records.models.Record.objects.filter(id=record_id).update(deadline=new_deadline.deadline)
+        ret = records.models.Record.objects.filter(id=record_id, owner_token=request.auth.owner_token).update(deadline=new_deadline.deadline)
         print(ret)
         if ret == 0:
             raise ValueError('Updated zero elements')
@@ -71,10 +74,10 @@ def update_deadline(request, record_id: int, new_deadline: cal.schemas.Deadline)
     else:
         return {'result': 'success', 'code': 0}
 
-@router.patch('/make_calendar')
+@router.patch('/make_calendar', auth=AuthBearer())
 def calendar_crate_record(request, id: int, additional: cal.schemas.CalendarRecordFromCrate):
     try:
-        r = records.models.Record.objects.filter(id=id)
+        r = records.models.Record.objects.filter(id=id, owner_token=request.auth.owner_token)
         print(r)
         if not r:
             raise ValueError('No such record')
